+++ tchannel/schemes/thrift.py
@@ -5,6 +5,7 @@ from __future__ import (
 from tornado import gen
 
 from . import THRIFT
+from tchannel.errors import ValueExpectedError
 from tchannel.thrift import serializer
 
 
@@ -58,6 +59,16 @@ class ThriftArgScheme(object):
 
         # success - non-void
         if len(result_spec) >= 1 and result_spec[0] is not None:
+
+            # value expected, but got none
+            # TODO - server side should use this same logic
+            if body.success is None:
+                raise ValueExpectedError(
+                    'Expected a value to be returned for %s, '
+                    'but recieved None - only void procedures can'
+                    'return None.' % request.endpoint
+                )
+
             response.body = body.success
             raise gen.Return(response)